Update python-jsonrpclib to 0.4.1 and remove python2 support
Summary:
Python2 support isn't used by any of existing packages in our repo.python2 support was re-added for for `lbryum` which is deprecated.
For now this is a dependendy only for `electrum` which is built with python3.

**Changelog:**

- Fixed a size computation issue in the request handler
- Added back support of Unix sockets on both server and client side.
- Fixed the CGI request handler
- Fixed the request handler wrapping on server side
- Fixed a memory leak in the Thread Pool, causing the `PooledJSONRPCServer` to crash after some uptime.

Depends on D9152

Signed-off-by: Algent Albrahimi <algent@protonmail.com>

Test Plan: Run `electrum`, create an account and manage the wallet.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9153